+++ PowerThreadPool/Core/Worker.cs
@@ -208,7 +208,7 @@ namespace PowerThreadPool
             try
             {
                 Interlocked.Increment(ref _powerPool._startCount);
-                Interlocked.Add(ref _powerPool._queueTime, (int)(runDateTime - Work.QueueDateTime).TotalMilliseconds);
+                Interlocked.Add(ref _powerPool._queueTime, (long)(runDateTime - Work.QueueDateTime).TotalMilliseconds);
                 object result = Work.Execute();
                 executeResult = Work.SetExecuteResult(result, null, Status.Succeed);
                 executeResult.StartDateTime = runDateTime;
